requirements.txt
streamlit==1.35.0
langchain==0.2.5
langchain-openai==0.1.9
langchain-anthropic==0.1.15
langchain-community==0.2.5
langchain-chroma==0.1.2
chromadb==0.5.0
anthropic==0.28.0
openai==1.35.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.2
geopy==2.4.1
folium==0.17.0
streamlit-folium==0.20.0
pytest==8.2.2
